run lint command issues fixed
