+++ client/src/components/HomeFAQ.jsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { motion, AnimatePresence } from "framer-motion";
+import {AnimatePresence } from "framer-motion";
 
 const HomeFAQ = ({ faqs }) => {
   const [openIndex, setOpenIndex] = useState(null);
